+++ src/instances/entities/Entities.java
@@ -46,7 +46,7 @@ public abstract class Entities extends Objects{
 
     
     @Override
-    public abstract void update(double deltaTime);
+    public abstract void update(float deltaTime);
 
     @Override
     public abstract void render(Graphics2D g2d);
